+++ src/main/java/ramdan/file/line/token/callback/Callback.java
@@ -1,4 +1,4 @@
-package ramdan.file.line.token.handler;
+package ramdan.file.line.token.callback;
 
 public interface Callback <T> {
     void  call(T t);
