+++ src/main/java/ramdan/file/line/token/callback/CallbackChain.java
@@ -1,4 +1,6 @@
-package ramdan.file.line.token.handler;
+package ramdan.file.line.token.callback;
+
+import ramdan.file.line.token.callback.Callback;
 
 public interface CallbackChain <S,T> extends Callback<S> {
     void setNext(Callback<T> next);
