+++ src/main/java/ramdan/file/line/token/callback/DefaultCallback.java
@@ -1,0 +1,12 @@
+package ramdan.file.line.token.callback;
+
+import ramdan.file.line.token.Tokens;
+import ramdan.file.line.token.callback.Callback;
+
+public class DefaultCallback implements Callback<Tokens> {
+
+    @Override
+    public void call(Tokens tokens) {
+
+    }
+}
