--- src/main/java/ramdan/file/line/token/handler/DefaultCallback.java
@@ -1,11 +1,0 @@
-package ramdan.file.line.token.handler;
-
-import ramdan.file.line.token.Tokens;
-
-public class DefaultCallback implements Callback<Tokens>{
-
-    @Override
-    public void call(Tokens tokens) {
-
-    }
-}
