+++ src/main/java/ramdan/file/line/token/handler/TokensHandler.java
@@ -1,8 +1,9 @@
 package ramdan.file.line.token.handler;
 
 import ramdan.file.line.token.Tokens;
+import ramdan.file.line.token.callback.Callback;
 
-public interface TokensHandler extends Handler,Callback<Tokens>{
+public interface TokensHandler extends Handler, Callback<Tokens> {
     void setNext(Callback<Tokens> next);
     @Deprecated
     Tokens process(Tokens tokes);
