+++ src/main/java/ramdan/file/line/token/listener/LineTokenCallbackLineListener.java
@@ -1,12 +1,9 @@
 package ramdan.file.line.token.listener;
 
-import jdk.nashorn.internal.parser.Token;
 import ramdan.file.line.token.Line;
 import ramdan.file.line.token.LineToken;
-import ramdan.file.line.token.Tokens;
 import ramdan.file.line.token.data.LineTokenData;
-import ramdan.file.line.token.handler.Callback;
-import ramdan.file.line.token.handler.LineTokenHandler;
+import ramdan.file.line.token.callback.Callback;
 
 public class LineTokenCallbackLineListener implements LineListener {
     private Callback<LineToken> handler;
